requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
black==25.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.1.2
click==8.3.0
contourpy==1.3.3
coverage==7.11.0
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
executing==2.2.1
fastapi==0.119.0
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
kaleido==1.1.0
kiwisolver==1.4.9
logfire==4.13.2
logistro==2.0.0
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.8.0
numpy==2.3.4
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
postgrest==2.22.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
realtime==2.22.0
redis==6.4.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
